+++ cmd/hg/io.go
@@ -16,12 +16,27 @@ func loadGraph(filename string) (*hypergraph.Hypergraph[string], error) {
 	return hypergraph.LoadJSON[string](f)
 }
 
-// saveGraph saves a hypergraph to a JSON file.
+// saveGraph saves a hypergraph to a JSON file atomically.
+// It writes to a temp file first, then renames to the target.
 func saveGraph(hg *hypergraph.Hypergraph[string], filename string) error {
-	f, err := os.Create(filename)
+	// Write to temp file first
+	tmpFile := filename + ".tmp"
+	f, err := os.Create(tmpFile)
 	if err != nil {
 		return err
 	}
-	defer f.Close()
-	return hg.SaveJSON(f)
+
+	if err := hg.SaveJSON(f); err != nil {
+		f.Close()
+		os.Remove(tmpFile)
+		return err
+	}
+
+	if err := f.Close(); err != nil {
+		os.Remove(tmpFile)
+		return err
+	}
+
+	// Atomic rename
+	return os.Rename(tmpFile, filename)
 }